#742 Durch E-Mail-Adress-Eingabe keine Informationen an Unbekannte weitergeben

+++ 01newsletter.php
@@ -79,207 +79,143 @@ else $echo_css = "";
 // Main_Top einfügen
 include($tempdir."main_top.html");
 
+// Infos einer ggf. übergebene E-Mail-Adresse aus Datenbank auslesen
+$row = _01newsletter_getEmailData($_REQUEST['email']);
 
 
+// Neuregistrierung oder Fake-Neuregistrierung
+if(isset($_REQUEST['sendregform']) && !empty($_REQUEST['sendregform']) &&
+	isset($_REQUEST['email']) && !empty($_REQUEST['email']) && check_mail($_REQUEST['email']) &&
+   ($settings['use_nutzungsbedingungen'] == 0 || $settings['use_nutzungsbedingungen'] == 1 && empty($settings['nutzungsbedingungen']) ||
+    isset($_REQUEST['ok_nutzungsbed']) && $_REQUEST['ok_nutzungsbed'] == 1)){
+	// E-Mail-Adresse wurde bereits registriert aber noch nicht aktiviert --> Eintrag löschen und neu registrieren
+    if(!empty($row['email']) && check_mail($row['email']) && strlen($row['acode']) == 32){
+    	$mysqli->query("DELETE FROM ".$mysql_tables['emailadds']." WHERE email = '".$mysqli->escape_string($row['email'])."' LIMIT 1");
+    	$row = FALSE;
+    }
 
+	// Echte Registrierung
+	if($row == FALSE){
+	    // Kategorien parsen:
+	    if(isset($_REQUEST['cats']) && !empty($_REQUEST['cats']) && is_array($_REQUEST['cats']) && !in_array("all",$_REQUEST['cats'])){
+	        $cats_string = ",";
+	        $cats_string .= implode(",",$_REQUEST['cats']);
+	        $cats_string .= ",";
+	        }
+	    else
+	        $cats_string = 0;
+	        
+	    $acode = md5(time().$_SERVER['REMOTE_ADDR'].mt_rand(1, 9999999999999).$_REQUEST['email']);
+	    
+	    $sql_insert = "INSERT INTO ".$mysql_tables['emailadds']." (acode,editcode,delcode,timestamp_reg,email,catids)
+	            VALUES(
+	               '".$acode."',
+	               '0',
+	               '0',
+	               '".time()."',
+	               '".$mysqli->escape_string($_REQUEST['email'])."',
+	               '".$mysqli->escape_string($cats_string)."'
+	               )";
+	    $mysqli->query($sql_insert) OR die($mysqli->error);
+	    
+	    $mail_inhalt = str_replace("#acodelink#",addParameter2Link($settings['formzieladdr'],"acode=".$acode),$lang['mailinhalt_acode']);
+	    $mail_inhalt = str_replace("#acode#",$acode,$mail_inhalt);
 
-// Wenn eine E-Mail-Adresse übergeben wurde
-if(isset($_REQUEST['email']) && !empty($_REQUEST['email']) && check_mail($_REQUEST['email'])){
-	// Überprüfen, ob E-Mail-Adresse bereits registriert ist
-	// Ja ->	Bearbeiten-Seite anzeigen (Account löschen / Kategorien ggf. ändern)
-	// Nein ->	E-Mail-Adresse neu registrieren bzw. Kategorie-Auswahl anzeigen
-	$list = $mysqli->query("SELECT * FROM ".$mysql_tables['emailadds']." WHERE email = '".$mysqli->escape_string($_REQUEST['email'])."' LIMIT 1");
-	
-	// E-Mail-Adresse bereits vorhanden?
-	if($list->num_rows > 0){
-		while($row = $list->fetch_assoc()){
-			// E-Mail-Adresse wurde bereits registriert aber noch nicht aktiviert
-			if(strlen($row['acode']) == 32){
-				// Neuen Aktivierungscode verschicken
-				if(isset($_REQUEST['action']) && $_REQUEST['action'] == "newacode"){
-					$mail_inhalt = str_replace("#acodelink#",addParameter2Link($settings['formzieladdr'],"acode=".$row['acode']),$lang['mailinhalt_acode']);
-					$mail_inhalt = str_replace("#acode#",$row['acode'],$mail_inhalt);
-					
-					$mail = new PHPMailer;
-					_01newsletter_configurePHPMailer($mail);
-					$mail->addAddress($row['email']);
-					$mail->Subject = $lang['mail_acode'];
-					$mail->Body    = $mail_inhalt;
-					$mail->send();
-					
-					$meldung = $lang['meldung_newacode'];
-					
-					include($tempdir."meldungen.html");
-					}
-				else{
-					$meldung = $lang['meldung_acode'];
-					$meldung .= "<br /><a href=\"".addParameter2Link($system_link,"action=newacode")."\">".$lang['resend_acode']."</a>";
-					
-					include($tempdir."meldungen.html");
-					}
-				
-				$formcodename = "acode";
-				include($tempdir."formular_acode.html");
-				}
-			// E-Mail-Adresse ist bereits registriert und aktiviert
-			else{
-				// Kategorien ändern
-				if(isset($_REQUEST['action']) && $_REQUEST['action'] == "update_account" && strlen($row['editcode']) == 32 && $row['editcode'] == $_POST['ecode']){
-					// Kategorien parsen:
-					if(isset($_REQUEST['cats']) && $_REQUEST['cats'] != "" && is_array($_REQUEST['cats']) && !in_array("all",$_REQUEST['cats'])){
-						$cats_string = ",";
-						$cats_string .= implode(",",$_REQUEST['cats']);
-						$cats_string .= ",";
-						}
-					else
-						$cats_string = 0;
-	
-					$mysqli->query("UPDATE ".$mysql_tables['emailadds']." SET catids = '".$mysqli->escape_string($cats_string)."', editcode='' WHERE email='".$mysqli->escape_string($row['email'])."' LIMIT 1");
-					
-					$meldung = $lang['meldung_changes'];
-					include($tempdir."meldungen.html");
-					}
-				// Lösch-Wunsch eintragen / Neuen Lösch-Code versenden
-				elseif(isset($_REQUEST['action']) && ($_REQUEST['action'] == "delabo" || $_REQUEST['action'] == "newdcode")){
-					$dcode = md5(time().$_SERVER['REMOTE_ADDR'].mt_rand(1, 9999999999999).$_REQUEST['email']);
-
-					$mysqli->query("UPDATE ".$mysql_tables['emailadds']." SET delcode='".$dcode."' WHERE email='".$mysqli->escape_string($row['email'])."' LIMIT 1");
-
-					$mail_inhalt = str_replace("#dcodelink#",addParameter2Link($settings['formzieladdr'],"dcode=".$dcode),$lang['mailinhalt_dcode']);
-					$mail_inhalt = str_replace("#dcode#",$dcode,$mail_inhalt);
-
-					$mail = new PHPMailer;
-					_01newsletter_configurePHPMailer($mail);
-					$mail->addAddress($row['email']);
-					$mail->Subject = $lang['mail_dcode'];
-					$mail->Body    = $mail_inhalt;
-					$mail->send();
-
-					$meldung = $lang['meldung_dcode_send'];
-					$formcodename = "dcode";
-
-					include($tempdir."meldungen.html");
-					include($tempdir."formular_acode.html");
-					}
-				// Eindeutigen Editier-Code zur Bearbeitung des eigenen Abonnements versenden
-				else{
-					$ecode = md5(time().$_SERVER['REMOTE_ADDR'].mt_rand(1, 9999999999999).$_REQUEST['email']);
-					
-					$mysqli->query("UPDATE ".$mysql_tables['emailadds']." SET editcode='".$ecode."' WHERE email='".$mysqli->escape_string($row['email'])."' LIMIT 1");
-					
-					$mail_inhalt = str_replace("#ecodelink#",addParameter2Link($settings['formzieladdr'],"ecode=".$ecode),$lang['mailinhalt_ecode']);
-					$mail_inhalt = str_replace("#ecode#",$ecode,$mail_inhalt);
-
-					$mail = new PHPMailer;
-					_01newsletter_configurePHPMailer($mail);
-					$mail->addAddress($row['email']);
-					$mail->Subject = $lang['mail_ecode'];
-					$mail->Body    = $mail_inhalt;
-					$mail->send();
-					
-					$meldung = $lang['meldung_ecode_send'];
-					$formcodename = "ecode";
-
-					include($tempdir."meldungen.html");
-					include($tempdir."formular_acode.html");
-					}
-				}
-			}
-		}
-	// E-Mail-Adresse noch nicht vorhanden -> Registrieren
-	else{
-		// Adresse direkt registrieren und Aktivierungsmail verschicken
-		if($settings['usecats'] == 0 && ($settings['use_nutzungsbedingungen'] == 0 || $settings['use_nutzungsbedingungen'] == 1 && empty($settings['nutzungsbedingungen']))){
-			$acode = md5(time().$_SERVER['REMOTE_ADDR'].mt_rand(1, 9999999999999).$_REQUEST['email']);
-			
-			$sql_insert = "INSERT INTO ".$mysql_tables['emailadds']." (acode,editcode,delcode,timestamp_reg,email,catids)
-				   		VALUES(
-						   '".$acode."',
-						   '0',
-						   '0',
-						   '".time()."',
-						   '".$mysqli->escape_string($_REQUEST['email'])."',
-						   '0'
-						   )";
-			$mysqli->query($sql_insert) OR die(mysql_error());
-			
-			$mail_inhalt = str_replace("#acodelink#",addParameter2Link($settings['formzieladdr'],"acode=".$acode),$lang['mailinhalt_acode']);
-			$mail_inhalt = str_replace("#acode#",$acode,$mail_inhalt);
-			
-			$mail = new PHPMailer;
-			_01newsletter_configurePHPMailer($mail);
-			$mail->addAddress($_REQUEST['email']);
-			$mail->Subject = $lang['mail_acode'];
-			$mail->Body    = $mail_inhalt;
-			$mail->send();
-			
-			$meldung = $lang['meldung_registriert'];
-			$meldung .= "<br /><a href=\"".addParameter2Link($system_link,"action=newacode")."\">".$lang['resend_acode']."</a>";
-
-			include($tempdir."meldungen.html");
-			$formcodename = "acode";
-			include($tempdir."formular_acode.html");
-			}
-		// Kategorien anzeigen
-		else{
-			// Daten in Datenbank speichern
-			if(isset($_REQUEST['sendregform']) && !empty($_REQUEST['sendregform']) &&
-			   ($settings['use_nutzungsbedingungen'] == 0 || $settings['use_nutzungsbedingungen'] == 1 && empty($settings['nutzungsbedingungen']) ||
-			    isset($_REQUEST['ok_nutzungsbed']) && $_REQUEST['ok_nutzungsbed'] == 1)){
-				// Kategorien parsen:
-				if(isset($_REQUEST['cats']) && !empty($_REQUEST['cats']) && is_array($_REQUEST['cats']) && !in_array("all",$_REQUEST['cats'])){
-					$cats_string = ",";
-					$cats_string .= implode(",",$_REQUEST['cats']);
-					$cats_string .= ",";
-					}
-				else
-					$cats_string = 0;
-					
-				$acode = md5(time().$_SERVER['REMOTE_ADDR'].mt_rand(1, 9999999999999).$_REQUEST['email']);
-				
-				$sql_insert = "INSERT INTO ".$mysql_tables['emailadds']." (acode,editcode,delcode,timestamp_reg,email,catids)
-				   		VALUES(
-						   '".$acode."',
-						   '0',
-						   '0',
-						   '".time()."',
-						   '".$mysqli->escape_string($_REQUEST['email'])."',
-						   '".$mysqli->escape_string($cats_string)."'
-						   )";
-				$mysqli->query($sql_insert) OR die($mysqli->error);
-				
-				$mail_inhalt = str_replace("#acodelink#",addParameter2Link($settings['formzieladdr'],"acode=".$acode),$lang['mailinhalt_acode']);
-				$mail_inhalt = str_replace("#acode#",$acode,$mail_inhalt);
-
-				$mail = new PHPMailer;
-				_01newsletter_configurePHPMailer($mail);
-				$mail->addAddress($_REQUEST['email']);
-				$mail->Subject = $lang['mail_acode'];
-				$mail->Body    = $mail_inhalt;
-				$mail->send();
-				
-				$meldung = $lang['meldung_registriert'];
-				$meldung .= "<br /><a href=\"".addParameter2Link($system_link,"action=newacode")."\">".$lang['resend_acode']."</a>";
-	
-				include($tempdir."meldungen.html");
-				$formcodename = "acode";
-				include($tempdir."formular_acode.html");
-				}
-			// Formular ausgeben
-			else{
-				$mailcats = "<option value=\"all\" selected=\"selected\">".$lang['allcats']."</option>\n";
-				
-				$list = $mysqli->query("SELECT * FROM ".$mysql_tables['mailcats']." ORDER BY catname");
-				while($row = $list->fetch_assoc()){
-					$mailcats .= "<option value=\"".$row['id']."\">".htmlentities($row['catname'],$htmlent_flags,$htmlent_encoding_acp)."</option>\n";
-					}
-				
-				include($tempdir."formular_registrierung.html");
-				}
-			}
-		}
-	
+	    $mail = new PHPMailer;
+	    _01newsletter_configurePHPMailer($mail);
+	    $mail->addAddress($_REQUEST['email']);
+	    $mail->Subject = $lang['mail_acode'];
+	    $mail->Body    = $mail_inhalt;
+	    $mail->send();
 	}
+    
+    $meldung = $lang['meldung_registriert'];
+
+    include($tempdir."meldungen.html");
+    $formcodename = "acode";
+    include($tempdir."formular_acode.html");
+    }
+// Registrierungsanfrage
+elseif(isset($_REQUEST['action']) && $_REQUEST['action'] == "add" || isset($_REQUEST['sendregform']) && !empty($_REQUEST['sendregform'])){
+	$mailcats = "<option value=\"all\" selected=\"selected\">".$lang['allcats']."</option>\n";
+
+	$list = $mysqli->query("SELECT * FROM ".$mysql_tables['mailcats']." ORDER BY catname");
+	while($row = $list->fetch_assoc()){
+		$mailcats .= "<option value=\"".$row['id']."\">".htmlentities($row['catname'],$htmlent_flags,$htmlent_encoding_acp)."</option>\n";
+	}
+
+	include($tempdir."formular_registrierung.html");
+}
+
+
+// Abonnement-Daten in DB ändern
+elseif(isset($_REQUEST['action']) && $_REQUEST['action'] == "update_account" && strlen($row['editcode']) == 32 && $row['editcode'] == $_POST['ecode']){
+	// Kategorien parsen:
+	if(isset($_REQUEST['cats']) && $_REQUEST['cats'] != "" && is_array($_REQUEST['cats']) && !in_array("all",$_REQUEST['cats'])){
+		$cats_string = ",";
+		$cats_string .= implode(",",$_REQUEST['cats']);
+		$cats_string .= ",";
+		}
+	else
+		$cats_string = 0;
+
+	$mysqli->query("UPDATE ".$mysql_tables['emailadds']." SET catids = '".$mysqli->escape_string($cats_string)."', editcode='' WHERE email='".$mysqli->escape_string($row['email'])."' LIMIT 1");
+	
+	$meldung = $lang['meldung_changes'];
+	include($tempdir."meldungen.html");
+	}
+// Eindeutigen Editier-Code zur Bearbeitung des eigenen Abonnements versenden
+elseif(isset($_REQUEST['action']) && $_REQUEST['action'] == "edit" || isset($_REQUEST['action']) && $_REQUEST['action'] == "update_account"){
+	if(!empty($row['email']) && check_mail($row['email'])){
+	    $ecode = md5(time().$_SERVER['REMOTE_ADDR'].mt_rand(1, 9999999999999).$_REQUEST['email']);
+	    
+	    $mysqli->query("UPDATE ".$mysql_tables['emailadds']." SET editcode='".$ecode."' WHERE email='".$mysqli->escape_string($row['email'])."' LIMIT 1");
+	    
+	    $mail_inhalt = str_replace("#ecodelink#",addParameter2Link($settings['formzieladdr'],"ecode=".$ecode),$lang['mailinhalt_ecode']);
+	    $mail_inhalt = str_replace("#ecode#",$ecode,$mail_inhalt);
+
+	    $mail = new PHPMailer;
+	    _01newsletter_configurePHPMailer($mail);
+	    $mail->addAddress($row['email']);
+	    $mail->Subject = $lang['mail_ecode'];
+	    $mail->Body    = $mail_inhalt;
+	    $mail->send();
+	}
+    
+    $meldung = $lang['meldung_ecode_send'];
+    $formcodename = "ecode";
+
+    include($tempdir."meldungen.html");
+    include($tempdir."formular_acode.html");
+    }
+
+
+// Lösch-Wunsch eintragen / Neuen Lösch-Code versenden
+elseif(isset($_REQUEST['action']) && $_REQUEST['action'] == "delabo"){
+	if(!empty($row['email']) && check_mail($row['email'])){
+		$dcode = md5(time().$_SERVER['REMOTE_ADDR'].mt_rand(1, 9999999999999).$_REQUEST['email']);
+
+		$mysqli->query("UPDATE ".$mysql_tables['emailadds']." SET delcode='".$dcode."' WHERE email='".$mysqli->escape_string($row['email'])."' LIMIT 1");
+
+		$mail_inhalt = str_replace("#dcodelink#",addParameter2Link($settings['formzieladdr'],"dcode=".$dcode),$lang['mailinhalt_dcode']);
+		$mail_inhalt = str_replace("#dcode#",$dcode,$mail_inhalt);
+
+		$mail = new PHPMailer;
+		_01newsletter_configurePHPMailer($mail);
+		$mail->addAddress($row['email']);
+		$mail->Subject = $lang['mail_dcode'];
+		$mail->Body    = $mail_inhalt;
+		$mail->send();
+	}
+
+	$meldung = $lang['meldung_dcode_send'];
+	$formcodename = "dcode";
+
+	include($tempdir."meldungen.html");
+	include($tempdir."formular_acode.html");
+	}
+
+
 // Aktivierungscode übergeben?
 elseif(isset($_REQUEST['acode']) && !empty($_REQUEST['acode']) && strlen($_REQUEST['acode']) == 32){
 	if($settings['send_benachrichtigung'] == 1){
@@ -306,14 +242,12 @@ elseif(isset($_REQUEST['acode']) && !empty($_REQUEST['acode']) && strlen($_REQUE
 		}
 	else{
 		$meldung = $lang['meldung_wrongacode'];
-		$formaction = "newacode";
-		$formcodename = "acode";
 
 		include($tempdir."meldungen.html");
-		include($tempdir."formular_resendcode.html");
-		include($tempdir."formular_acode.html");
 		}
 	}
+
+
 // Bearbeitungscode übergeben?
 elseif(isset($_REQUEST['ecode']) && !empty($_REQUEST['ecode']) && strlen($_REQUEST['ecode']) == 32){
 	$list = $mysqli->query("SELECT * FROM ".$mysql_tables['emailadds']." WHERE editcode = '".$mysqli->escape_string($_REQUEST['ecode'])."' LIMIT 1");
@@ -349,7 +283,7 @@ elseif(isset($_REQUEST['ecode']) && !empty($_REQUEST['ecode']) && strlen($_REQUE
 		}
 	}
 	else{
-		$meldung = $lang['meldung_falschemail'];
+		$meldung = $lang['meldung_wrongacode'];
 
 		include($tempdir."meldungen.html");
 	}
@@ -357,10 +291,12 @@ elseif(isset($_REQUEST['ecode']) && !empty($_REQUEST['ecode']) && strlen($_REQUE
 // Löschcode übergeben?
 elseif(isset($_REQUEST['dcode']) && !empty($_REQUEST['dcode']) && strlen($_REQUEST['dcode']) == 32){
 	$list = $mysqli->query("SELECT email FROM ".$mysql_tables['emailadds']." WHERE delcode='".$mysqli->escape_string($_REQUEST['dcode'])."' AND delcode != '0' LIMIT 1");
-	$row_email = $list->fetch_assoc();
-	
-	$mysqli->query("DELETE FROM ".$mysql_tables['temp_table']." WHERE email = '".$mysqli->escape_string($row_email['email'])."' AND email != ''");
-	$mysqli->query("DELETE FROM ".$mysql_tables['emailadds']." WHERE delcode='".$mysqli->escape_string($_REQUEST['dcode'])."' AND delcode != '0' LIMIT 1");
+	if($list->num_rows == 1){
+		$row_email = $list->fetch_assoc();
+		
+		$mysqli->query("DELETE FROM ".$mysql_tables['temp_table']." WHERE email = '".$mysqli->escape_string($row_email['email'])."' AND email != ''");
+		$mysqli->query("DELETE FROM ".$mysql_tables['emailadds']." WHERE delcode='".$mysqli->escape_string($_REQUEST['dcode'])."' AND delcode != '0' LIMIT 1");
+	}
 
 	if($mysqli->affected_rows == 1){
 		$meldung = $lang['meldung_deleted'];
@@ -368,20 +304,9 @@ elseif(isset($_REQUEST['dcode']) && !empty($_REQUEST['dcode']) && strlen($_REQUE
 	}
 	else{
 		$meldung = $lang['meldung_wrongacode'];
-		$formaction = "newdcode";
-		$formcodename = "dcode";
 
 		include($tempdir."meldungen.html");
-		include($tempdir."formular_resendcode.html");
-		include($tempdir."formular_acode.html");
 	}
-}
-// Fehlermeldung: Fehlerhafte E-Mail-Adresse eingegeben anzeigen
-elseif(isset($_REQUEST['email']) && !empty($_REQUEST['email'])){
-	$meldung = $lang['meldung_falschemail'];
-
-	include($tempdir."meldungen.html");
-	include($tempdir."formular_addemail.html");
 }
 // Nichts übergeben -> Register-Formular anzeigen
 else
